kola/tests/etcd: register discovery tests in discovery file
