+++ kola/tests/etcd/util.go
@@ -24,44 +24,9 @@ import (
 	"strings"
 	"time"
 
-	"github.com/coreos/mantle/kola/register"
 	"github.com/coreos/mantle/platform"
 	"github.com/coreos/mantle/util"
 )
-
-func init() {
-	// test etcd discovery with 0.4.7
-	register.Register(&register.Test{
-		Run:         DiscoveryV1,
-		Manual:      true,
-		ClusterSize: 3,
-		Name:        "coreos.etcd0.discovery",
-		UserData: `#cloud-config
-coreos:
-  etcd:
-    name: $name
-    discovery: $discovery
-    addr: $private_ipv4:2379
-    peer-addr: $private_ipv4:2380`,
-	})
-
-	// test etcd discovery with 2.0 with new cloud config
-	register.Register(&register.Test{
-		Run:         DiscoveryV2,
-		ClusterSize: 3,
-		Name:        "coreos.etcd2.discovery",
-		UserData: `#cloud-config
-
-coreos:
-  etcd2:
-    name: $name
-    discovery: $discovery
-    advertise-client-urls: http://$private_ipv4:2379
-    initial-advertise-peer-urls: http://$private_ipv4:2380
-    listen-client-urls: http://0.0.0.0:2379,http://0.0.0.0:4001
-    listen-peer-urls: http://$private_ipv4:2380,http://$private_ipv4:7001`,
-	})
-}
 
 // GetClusterHealth polls etcdctl cluster-health command until success
 // or maximum retries have been reached. Can be effectively used to
